Edited further
