+++ safefoodie_login_merge/.dart_tool/package_config_subset
@@ -114,6 +114,10 @@ provider
 2.12
 file:///C:/Downloaded%20Utilities/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/
 file:///C:/Downloaded%20Utilities/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/lib/
+settings_ui
+2.12
+file:///C:/Downloaded%20Utilities/flutter/.pub-cache/hosted/pub.dartlang.org/settings_ui-2.0.2/
+file:///C:/Downloaded%20Utilities/flutter/.pub-cache/hosted/pub.dartlang.org/settings_ui-2.0.2/lib/
 source_span
 2.14
 file:///C:/Downloaded%20Utilities/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.9.0/
