+++ safefoodie_login_merge/lib/LoginPage.dart
@@ -132,123 +132,5 @@ class _LoginScreenState extends State<LoginScreen> {
   }
 }
 //======================================================================================
-//======================================================================================
-//Firebase Authorization
 
-//======================================================================================
-//VIEW LISTS PAGE START
-//======================================================================================
-class _ViewListsState extends State<ViewListsPage> {
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-        appBar: AppBar(title: const Text('Account recovery')),
-        body: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Container(
-              alignment: Alignment.center,
-              child: Stack(
-                children: <Widget>[
-                  Container(
-                    padding: EdgeInsets.fromLTRB(15, 110, 0, 0),
-                    child:
-                        Text("Password reset", style: TextStyle(fontSize: 40)),
-                  )
-                ],
-              ),
-            ),
-//======================================
-//Text field
-            Container(
-              padding: EdgeInsets.only(top: 35, left: 20, right: 30),
-              child: Column(
-                children: <Widget>[
-                  TextField(
-                    // controller: _emailController,
-                    decoration: const InputDecoration(
-                      enabledBorder: const OutlineInputBorder(
-                        borderSide:
-                            const BorderSide(color: Colors.white, width: 0.0),
-                      ),
-                      border: const OutlineInputBorder(),
-                      labelStyle: TextStyle(color: Colors.white),
-                      labelText: 'Enter account email',
-                    ),
-                  ),
-                  SizedBox(
-                    height: 20,
-                  ),
-                  SizedBox(
-                    height: 5.0,
-                  ),
-//======================================
-// sign up button
-                  Container(
-                    height: 40,
-                    padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
-                    child: Material(
-                      borderRadius: BorderRadius.circular(20),
-                      shadowColor: Colors.greenAccent,
-                      color: Colors.green,
-                      elevation: 7,
-                      child: GestureDetector(
-                          onTap: () async {
-                            //      _register();
-                          },
-                          child: Center(
-                              child: Text('SEND RESET EMAIL',
-                                  style: TextStyle(
-                                      color: Colors.white,
-                                      fontWeight: FontWeight.bold,
-                                      fontFamily: 'Montserrat')))),
-                    ),
-                  ),
-                  SizedBox(
-                    height: 15,
-                  ),
-//======================================
-//back button
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      InkWell(
-                        onTap: () {
-                          Navigator.of(context).pop();
-                        },
-                        child: Text('Go Back',
-                            style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'Montserrat',
-                                fontWeight: FontWeight.bold,
-                                decoration: TextDecoration.underline)),
-                      )
-                    ],
-                  )
-                ],
-              ),
-            )
-          ],
-        ));
-  }
 
-// Need to import these after Viewlist page is Created //
-
-//======================================================================================
-//NAV FUNCTIONS
-//======================================================================================
-//Nav function for view lists page
-  void _ToViewListsScreen(BuildContext context) {
-    Navigator.of(context)
-        .push(MaterialPageRoute(builder: (context) => ViewListsPage()));
-  }
-}
-
-//======================================================================================
-//PAGE CLASS INITIALIZATION
-//======================================================================================
-//View lists page
-class ViewListsPage extends StatefulWidget {
-  @override
-  _ViewListsState createState() => _ViewListsState();
-}
