+++ safefoodie_login_merge/lib/Search.dart
@@ -1,0 +1,107 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'Mainpage.dart';
+
+final FirebaseAuth _auth = FirebaseAuth.instance;
+
+class SearchPage extends StatefulWidget {
+  @override
+  State<SearchPage> createState() => _SearchPageState();
+}
+
+class _SearchPageState extends State<SearchPage> {
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+        appBar: AppBar(title: const Text('Search Page')),
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Container(
+              alignment: Alignment.center,
+              child: Stack(
+                children: <Widget>[
+                  Container(
+                    padding: EdgeInsets.fromLTRB(15, 110, 0, 0),
+                    child:
+                        Text("Password reset", style: TextStyle(fontSize: 40)),
+                  )
+                ],
+              ),
+            ),
+//======================================
+//Text field
+            Container(
+              padding: EdgeInsets.only(top: 35, left: 20, right: 30),
+              child: Column(
+                children: <Widget>[
+                  TextField(
+                    // controller: _emailController,
+                    decoration: const InputDecoration(
+                      enabledBorder: const OutlineInputBorder(
+                        borderSide:
+                            const BorderSide(color: Colors.white, width: 0.0),
+                      ),
+                      border: const OutlineInputBorder(),
+                      labelStyle: TextStyle(color: Colors.white),
+                      labelText: 'Enter account email',
+                    ),
+                  ),
+                  SizedBox(
+                    height: 20,
+                  ),
+                  SizedBox(
+                    height: 5.0,
+                  ),
+//======================================
+// sign up button
+                  Container(
+                    height: 40,
+                    padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
+                    child: Material(
+                      borderRadius: BorderRadius.circular(20),
+                      shadowColor: Colors.greenAccent,
+                      color: Colors.green,
+                      elevation: 7,
+                      child: GestureDetector(
+                          onTap: () async {
+                            //      _register();
+                          },
+                          child: Center(
+                              child: Text('SEND RESET EMAIL',
+                                  style: TextStyle(
+                                      color: Colors.white,
+                                      fontWeight: FontWeight.bold,
+                                      fontFamily: 'Montserrat')))),
+                    ),
+                  ),
+                  SizedBox(
+                    height: 15,
+                  ),
+//======================================
+//back button
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      InkWell(
+                        onTap: () {
+                          Navigator.of(context).pop();
+                        },
+                        child: Text('Go Back',
+                            style: TextStyle(
+                                color: Colors.white,
+                                fontFamily: 'Montserrat',
+                                fontWeight: FontWeight.bold,
+                                decoration: TextDecoration.underline)),
+                      )
+                    ],
+                  )
+                ],
+              ),
+            )
+          ],
+        ));
+  }
+}
