+++ safefoodie_login_merge/pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   firebase_auth: ^3.11.2
   cloud_firestore: ^3.5.0
   provider: ^6.0.3
+  settings_ui: ^2.0.2
 
 
 dev_dependencies:
